requirements.txt
ipykernel
Faker
numpy
google-cloud-storage
psycopg2
fastapi
uvicorn[standard]